--- connector_ecommerce/delivery.py
@@ -1,30 +1,0 @@
-# -*- encoding: utf-8 -*-
-##############################################################################
-#
-#    Author Guewen Baconnier. Copyright 2011 Camptocamp SA
-#
-#    This program is free software: you can redistribute it and/or modify
-#    it under the terms of the GNU General Public License as published by
-#    the Free Software Foundation, either version 3 of the License, or
-#    (at your option) any later version.
-#
-#    This program is distributed in the hope that it will be useful,
-#    but WITHOUT ANY WARRANTY; without even the implied warranty of
-#    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-#    GNU General Public License for more details.
-#
-#    You should have received a copy of the GNU General Public License
-#    along with this program.  If not, see <http://www.gnu.org/licenses/>.
-#
-##############################################################################
-
-from openerp.osv.orm import Model
-from openerp.osv import fields
-
-class delivery_carrier(Model):
-    _inherit = "delivery.carrier"
-
-    _columns = {
-        # TODO to be removed once the job export tracking is there
-        'export_needs_tracking': fields.boolean('Export only shippings with a tracking number'),
-    }
