--- connector_ecommerce/delivery_view.xml
@@ -1,19 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<openerp>
-    <data>
-
-        <record id="view_delivery_carrier_form_inherit" model="ir.ui.view">
-            <field name="name">delivery.carrier.form.inherit</field>
-            <field name="model">delivery.carrier</field>
-            <field name="inherit_id" ref="delivery.view_delivery_carrier_form" />
-            <field name="arch" type="xml">
-                 <data>
-                    <field name="product_id" position="after">
-                        <field name="export_needs_tracking"/>
-                    </field>
-                </data>
-            </field>
-        </record>
-
-    </data>
-</openerp>
