+++ connector_ecommerce/invoice_view.xml
@@ -2,5 +2,19 @@
 <openerp>
     <data>
 
+        <record id="view_invoice_connector_form" model="ir.ui.view">
+            <field name="name">account.invoice.connector.form</field>
+            <field name="model">account.invoice</field>
+            <field name="inherit_id" ref="account.invoice_form"/>
+            <field name="arch" type="xml">
+                <page string="Payments" position="after">
+                    <!-- change the invisible attribute to 0 when used
+                         in submodules -->
+                    <page name="connector" string="Connectors" invisible="1">
+                    </page>
+                </page>
+            </field>
+        </record>
+
     </data>
 </openerp>
