+++ connector_ecommerce/old_xml_stuff.xml
@@ -1,11 +1,5 @@
 <?xml version="1.0" encoding="utf-8"?>
 
-<!--
-    Copyright (C) 2012 Akretion (http://www.akretion.com/)
-    @author Alexis de Lattre <alexis.delattre@akretion.com>
-    The licence is in the file __openerp__.py
--->
-
 <openerp>
 <data>
 
@@ -26,3 +20,4 @@
 
 </data>
 </openerp>
+
