+++ connector_ecommerce/product.py
@@ -1,170 +1,175 @@
-    # -*- encoding: utf-8 -*-
-###############################################################################
-#                                                                             #
-#   connector_ecommerce for OpenERP                                       #
-#   Copyright (C) 2011 Akretion Sébastien BEAU <sebastien.beau@akretion.com>  #
-#                                                                             #
-#   This program is free software: you can redistribute it and/or modify      #
-#   it under the terms of the GNU Affero General Public License as            #
-#   published by the Free Software Foundation, either version 3 of the        #
-#   License, or (at your option) any later version.                           #
-#                                                                             #
-#   This program is distributed in the hope that it will be useful,           #
-#   but WITHOUT ANY WARRANTY; without even the implied warranty of            #
-#   MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the             #
-#   GNU Affero General Public License for more details.                       #
-#                                                                             #
-#   You should have received a copy of the GNU Affero General Public License  #
-#   along with this program.  If not, see <http://www.gnu.org/licenses/>.     #
-#                                                                             #
-###############################################################################
+# -*- coding: utf-8 -*-
+##############################################################################
+#
+#    Author: Sébastien BEAU
+#    Copyright 2011-2013 Akretion
+#
+#    This program is free software: you can redistribute it and/or modify
+#    it under the terms of the GNU Affero General Public License as
+#    published by the Free Software Foundation, either version 3 of the
+#    License, or (at your option) any later version.
+#
+#    This program is distributed in the hope that it will be useful,
+#    but WITHOUT ANY WARRANTY; without even the implied warranty of
+#    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+#    GNU Affero General Public License for more details.
+#
+#    You should have received a copy of the GNU Affero General Public License
+#    along with this program.  If not, see <http://www.gnu.org/licenses/>.
+#
+##############################################################################
+
+from openerp.osv import orm, fields
+from openerp.addons.connector.session import ConnectorSession
+from .event import on_product_price_changed
 
 
-from openerp.osv.orm import Model
-from openerp.osv import fields
-
-from openerp.addons.connector.decorator import only_for_referential
-from openerp.addons.connector.decorator import commit_now
-
-class product_product(Model):
-    _inherit='product.product'
-
-    def _check_if_export(self, cr, uid, external_session, product, context=None):
-        if context.get('export_product') == 'simple':
-            return True
-        return False
-
-    @only_for_referential(ref_categ ='Multichannel Sale')
-    def _get_last_exported_date(self, cr, uid, external_session, context=None):
-        shop = external_session.sync_from_object
-        if context.get('export_product') == 'simple':
-            return shop.last_products_export_date
-        elif context.get('export_product') == 'special':
-            return shop.last_special_products_export_date
-        return False
-
-    @only_for_referential(ref_categ ='Multichannel Sale')
-    @commit_now
-    def _set_last_exported_date(self, cr, uid, external_session, date, context=None):
-        shop = external_session.sync_from_object
-        if context.get('export_product') == 'simple':
-            return self.pool.get('sale.shop').write(cr, uid, shop.id, {'last_products_export_date': date}, context=context)
-        elif context.get('export_product') == 'special':
-            return self.pool.get('sale.shop').write(cr, uid, shop.id, {'last_special_products_export_date': date}, context=context)
-
-    @only_for_referential(ref_categ ='Multichannel Sale')
-    def get_ids_and_update_date(self, cr, uid, external_session, ids=None, last_exported_date=None, context=None):
-        res = (), {} # list of ids, dict of ids to date_changed
-        shop = external_session.sync_from_object
-        if shop.exportable_product_ids:
-            product_ids = [product.id for product in shop.exportable_product_ids if self._check_if_export(cr, uid, external_session, product, context=context)]
-            if ids:
-                product_ids = set(ids).intersection(set(product_ids))
-            if product_ids:
-                res = super(product_product, self).get_ids_and_update_date(cr, uid, external_session,
-                                                            ids=product_ids,
-                                                            last_exported_date=last_exported_date,
-                                                            context=context)
-        return res
-
-
-    def _get_categories_ids_for_shop(self, cr, uid, product_id, shop_id, context=None):
-        shop_categ_ids = self.pool.get('sale.shop').read(cr, uid, shop_id,
-                                ['exportable_category_ids'],
-                                context=context)['exportable_category_ids']
-        product = self.read(cr, uid, product_id, ['categ_ids', 'categ_id'], context=context)
-        product_categ_ids = product['categ_ids']
-        if product['categ_id'][0] not in product_categ_ids:
-            product_categ_ids.append(product['categ_id'][0])
-        res = []
-        for categ in product_categ_ids:
-            if categ in shop_categ_ids:
-                res.append(categ)
-        return res
-
-    def _get_categories_ids_for_shop(self, cr, uid, product_id, shop_id, context=None):
-        shop_obj = self.pool.get('sale.shop')
-        shop_values = shop_obj.read(cr, uid, shop_id,
-                                    ['exportable_category_ids'],
-                                    context=context)
-        shop_categ_ids = set(shop_values['exportable_category_ids'])
-        product = self.read(cr, uid, product_id, ['categ_ids', 'categ_id'], context=context)
-        product_categ_ids = set(product['categ_ids'])
-        product_categ_ids.add(product['categ_id'][0])
-        return list(prod_categ_ids & shop_categ_ids)
-
-    def _get_or_create_ext_category_ids_for_shop(self, cr, uid, external_session, product_id, context=None):
-        res = []
-        categ_obj = self.pool.get('product.category')
-        for oe_categ_id in self._get_categories_ids_for_shop(cr, uid, product_id, external_session.sync_from_object.id, context=context):
-            res.append(categ_obj.get_or_create_extid(cr, uid, external_session, oe_categ_id, context=context))
-        return res
-
-
-
-class product_template(Model):
+class product_template(orm.Model):
     _inherit = 'product.template'
 
     #TODO implement set function and also support multi tax
     def _get_tax_group_id(self, cr, uid, ids, field_name, args, context=None):
         result = {}
         for product in self.browse(cr, uid, ids, context=context):
-            result[product.id] = product.taxes_id and product.taxes_id[0].group_id.id
+            taxes = product.taxes_id
+            result[product.id] = taxes[0].group_id.id if taxes else False
         return result
 
     _columns = {
-        'tax_group_id': fields.function(_get_tax_group_id,
+        'tax_group_id': fields.function(
+                            _get_tax_group_id,
                             string='Tax Group',
                             type='many2one',
                             relation='account.tax.group',
                             store=False,
-                            help=('Tax group are use with some external',
-                                  ' system like magento or prestashop')),
+                            help='Tax group are used with some external '
+                                 'system like magento or prestashop'),
     }
 
+    def _price_changed(self, cr, uid, ids, vals, context=None):
+        """ Fire the ``on_product_price_changed`` on all the variants of
+        the template if the price if the product could have changed.
 
-class product_category(Model):
-    _inherit = "product.category"
+        If one of the field used in a sale pricelist item has been
+        modified, we consider that the price could have changed.
 
-    def collect_children(self, category, children=None):
-        if children is None:
-            children = []
+        There is no guarantee that's the price actually changed,
+        because it depends on the pricelists.
+        """
+        if context is None:
+            context = {}
+        type_obj = self.pool['product.price.type']
+        price_fields = type_obj.sale_price_fields(cr, uid, context=context)
+        # restrict the fields to the template ones only, so if
+        # the write has been done on product.product, we won't
+        # update all the variant if a price field of the
+        # variant has been changed
+        tmpl_fields = [field for field in vals if field in self._columns]
+        if any(field in price_fields for field in tmpl_fields):
+            product_obj = self.pool['product.product']
+            session = ConnectorSession(cr, uid, context=context)
+            product_ids = product_obj.search(cr, uid,
+                                             [('product_tmpl_id', 'in', ids)],
+                                             context=context)
+            # when the write is done on the product.product, avoid
+            # to fire the event 2 times
+            if context.get('from_product_ids'):
+                product_ids = list(set(product_ids) -
+                                   set(context['from_product_ids']))
+            for prod_id in product_ids:
+                on_product_price_changed.fire(session,
+                                              product_obj._name,
+                                              prod_id)
 
-        for child in category.child_id:
-            children.append(child.id)
-            self.collect_children(child, children)
+    def write(self, cr, uid, ids, vals, context=None):
+        result = super(product_template, self).write(cr, uid, ids,
+                                                     vals, context=context)
+        self._price_changed(cr, uid, ids, vals, context=context)
+        return result
 
-        return children
 
-    def _get_recursive_children_ids(self, cr, uid, ids, name, args, context=None):
-        res = {}
-        for category in self.browse(cr, uid, ids):
-            res[category.id] = self.collect_children(category, [category.id])
-        return res
+class product_product(orm.Model):
+    _inherit = 'product.product'
+
+    def _get_checkpoint(self, cr, uid, ids, name, arg, context=None):
+        result = {}
+        checkpoint_obj = self.pool.get('connector.checkpoint')
+        model_obj = self.pool.get('ir.model')
+        model_id = model_obj.search(cr, uid,
+                                    [('model', '=', 'product.product')],
+                                    context=context)[0]
+        for product_id in ids:
+            point_ids = checkpoint_obj.search(cr, uid,
+                                              [('model_id', '=', model_id),
+                                               ('record_id', '=', product_id),
+                                               ('state', '=', 'need_review')],
+                                              context=context)
+            result[product_id] = bool(point_ids)
+        return result
 
     _columns = {
-        'recursive_children_ids': fields.function(_get_recursive_children_ids, method=True, type='one2many', relation="product.category", string='All Child Categories'),
+        'has_checkpoint': fields.function(_get_checkpoint,
+                                          type='boolean',
+                                          readonly=True,
+                                          string='Has Checkpoint'),
     }
 
-    @only_for_referential(ref_categ ='Multichannel Sale')
-    def _get_last_exported_date(self, cr, uid, external_session, context=None):
-        shop = external_session.sync_from_object
-        return shop.last_category_export_date
+    def _price_changed(self, cr, uid, ids, vals, context=None):
+        """ Fire the ``on_product_price_changed`` if the price
+        if the product could have changed.
 
-    @only_for_referential(ref_categ ='Multichannel Sale')
-    @commit_now
-    def _set_last_exported_date(self, cr, uid, external_session, date, context=None):
-        shop = external_session.sync_from_object
-        return self.pool.get('sale.shop').write(cr, uid, shop.id, {'last_category_export_date': date}, context=context)
+        If one of the field used in a sale pricelist item has been
+        modified, we consider that the price could have changed.
 
-    def get_ids_and_update_date(self, cr, uid, external_session, ids=None, last_exported_date=None, context=None):
-        shop = external_session.sync_from_object
-        if shop.exportable_category_ids:
-            res = super(product_category, self).get_ids_and_update_date(cr, uid, external_session,
-                                                            ids=[categ.id for categ in shop.exportable_category_ids],
-                                                            last_exported_date=last_exported_date,
-                                                            context=context)
-        else:
-            res = (), {} # list of ids, dict of ids to date_changed
-        return res
+        There is no guarantee that's the price actually changed,
+        because it depends on the pricelists.
+        """
+        type_obj = self.pool['product.price.type']
+        price_fields = type_obj.sale_price_fields(cr, uid, context=context)
+        if any(field in price_fields for field in vals):
+            session = ConnectorSession(cr, uid, context=context)
+            for prod_id in ids:
+                on_product_price_changed.fire(session, self._name, prod_id)
 
+    def write(self, cr, uid, ids, vals, context=None):
+        if context is None:
+            context = {}
+        context = context.copy()
+        context['from_product_ids'] = ids
+        result = super(product_product, self).write(
+            cr, uid, ids, vals, context=context)
+        self._price_changed(cr, uid, ids, vals, context=context)
+        return result
+
+    def create(self, cr, uid, vals, context=None):
+        product_ids = super(product_product, self).create(
+            cr, uid, vals, context=None)
+        self._price_changed(cr, uid, [product_ids], vals, context=context)
+        return product_ids
+
+
+class product_price_type(orm.Model):
+    _inherit = 'product.price.type'
+
+    _columns = {
+        'pricelist_item_ids': fields.one2many(
+            'product.pricelist.item', 'base',
+            string='Pricelist Items',
+            readonly=True)
+    }
+
+    def sale_price_fields(self, cr, uid, context=None):
+        """ Returns a list of fields used by sale pricelists.
+        Used to know if the sale price could have changed
+        when one of these fields has changed.
+        """
+        item_obj = self.pool['product.pricelist.item']
+        item_ids = item_obj.search(
+            cr, uid,
+            [('price_version_id.pricelist_id.type', '=', 'sale')],
+            context=context)
+        type_ids = self.search(cr, uid,
+                               [('pricelist_item_ids', 'in', item_ids)],
+                               context=context)
+        types = self.read(cr, uid, type_ids, ['field'], context=context)
+        return [t['field'] for t in types]
