+++ connector_ecommerce/product_view.xml
@@ -10,6 +10,7 @@
 
 <!--This is a product inherited form that will be dynamically populated-->
 <record id="product_normal_form_view" model="ir.ui.view">
+   <field name="name">product.product.form</field>
    <field name="model">product.product</field>
    <field name="inherit_id" ref="product.product_normal_form_view" />
    <field name="arch" type="xml">
