+++ connector_ecommerce/security/ir.model.access.csv
@@ -1,18 +1,3 @@
 "id","name","model_id:id","group_id:id","perm_read","perm_write","perm_create","perm_unlink"
-"access_connector_ecommerce_external_shop_group_manager","connector_ecommerce_external_shop_group","model_external_shop_group","base.group_sale_manager",1,1,1,1
-"access_connector_external_referential","connector_external_referential","connector.model_external_referential","base.group_sale_manager",1,1,1,1
-"access_connector_external_referential_type","connector_external_referential_type","connector.model_external_referential_type","base.group_sale_manager",1,1,1,1
-"access_connector_external_referential_version","connector_external_referential_version","connector.model_external_referential_version","base.group_sale_manager",1,1,1,1
-"access_connector_external_mapping_template","connector_external_mapping_template","connector.model_external_mapping_template","base.group_sale_manager",1,1,1,1
-"access_connector_external_mappinglines_template","connector_external_mappinglines_template","connector.model_external_mappinglines_template","base.group_sale_manager",1,1,1,1
-"access_connector_external_referential","connector_external_referential","connector.model_external_referential","base.group_sale_manager",1,1,1,1
-"access_connector_external_mapping_line","connector_external_mapping_line","connector.model_external_mapping_line","base.group_sale_manager",1,1,1,1
-"access_connector_external_mapping","connector_external_mapping","connector.model_external_mapping","base.group_sale_manager",1,1,1,1
-"access_connector_external_report","connector_external_report","connector.model_external_report","base.group_sale_manager",1,1,1,1
-"access_connector_external_report_history","connector_external_report_history","connector.model_external_report_history","base.group_sale_manager",1,1,1,1
-"access_connector_external_report_line","connector_external_report_line","connector.model_external_report_line","base.group_sale_manager",1,1,1,1
-"access_sale_sale_shop","sale_sale_shop","sale.model_sale_shop","base.group_sale_manager",1,1,1,1
-"access_connector_ecommerce_ir_model_data","connector_ecommerce_ir_model_data","base.model_ir_model_data","base.group_sale_manager",1,1,1,1
-"access_connector_ecommerce_external_shop_group","connector_ecommerce_external_shop_group","model_external_shop_group","base.group_user",1,0,0,0
 "access_account_tax_group_user","Read-only access to account.tax.group","model_account_tax_group","base.group_user",1,0,0,0
 "access_account_tax_group_account_manager","RW access to account.tax.group","model_account_tax_group","account.group_account_manager",1,1,1,1
