--- connector_ecommerce/settings/external.referential.category.csv
@@ -1,2 +1,0 @@
-"id","name"
-"multichannels","Multichannel Sale"
