--- connector_ecommerce/wizard/import_order.xml
@@ -1,34 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<openerp>
-    <data>
-        <record model="ir.ui.view" id="sale_order_import_wizard_view">
-            <field name="name">sale.order.import.wizard.view</field>
-            <field name="model">sale.order.import.wizard</field>
-            <field name="arch" type="xml">
-                <form string="Import a specifique order from an external system">
-                    <field name="order_number" nolabel="1" colspan='4'/>
-                    <button icon="gtk-cancel" special="cancel" string="Cancel"/>
-                    <button icon="gtk-ok" name="import_order" string="Import Order" type="object"/>
-                </form>
-            </field>
-        </record>
-
-        <record id="action_import_order_wizard" model="ir.actions.act_window">
-            <field name="name">Import Order</field>
-            <field name="type">ir.actions.act_window</field>
-            <field name="res_model">sale.order.import.wizard</field>
-            <field name="view_type">form</field>
-            <field name="view_mode">form</field>
-            <field name="target">new</field>
-            <field name="context">{}</field>
-        </record>
-
-        <record id="ir_action_import_order_wizard" model="ir.values">
-            <field name="key2">client_action_multi</field>
-            <field name="model">sale.shop</field>
-            <field name="name">Import Order</field>
-            <field eval="'ir.actions.act_window,%d'%action_import_order_wizard" name="value"/>
-            <field eval="True" name="object"/>
-        </record>
-    </data>
-</openerp>
